Refactor to make way for multiple classes
